requirements.txt
alembic==1.14.1
bcrypt==4.3.0
blinker==1.8.2
cachetools==5.5.1
captcha==0.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Security-Too==5.3.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.10
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.40
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
